[x86][exceptions] Don't truncate CR2 when reporting to usermode
This would cause usermode accesses to kernel memory to have incorrect
CR2s in their crashlogger dump.

+++ kernel/arch/x86/include/arch/x86.h
@@ -37,7 +37,7 @@ enum handler_return platform_irq(x86_iframe_t *frame);
 struct arch_exception_context {
     bool is_page_fault;
     x86_iframe_t *frame;
-    uint32_t cr2;
+    uint64_t cr2;
 };
 
 struct x86_64_context_switch_frame {
